ansible/inventories/generate_inventory.py
#!/usr/bin/env python3
"""
Minimal Ansible inventory generator for the hyperscaler project.
This script creates basic inventory structure from cluster.yaml configuration.
"""

import yaml
import sys
from pathlib import Path


def generate_inventory(cluster_config_path: str = "cluster.yaml"):
    """Generate basic Ansible inventory structure."""
    print(f"Generating Ansible inventory from {cluster_config_path}")
    print("Inventory generation will be implemented here")

    # Placeholder for actual inventory generation logic
    inventory_structure = {
        "hpc_cluster": {
            "hosts": {},
            "vars": {}
        },
        "cloud_cluster": {
            "hosts": {},
            "vars": {}
        }
    }

    print("Basic inventory structure created")
    return inventory_structure


if __name__ == "__main__":
    config_path = sys.argv[1] if len(sys.argv) > 1 else "cluster.yaml"
    generate_inventory(config_path)
