ansible/requirements.txt
# Ansible Core and Dependencies
# Core Ansible installation
ansible>=8.0.0

# Python dependencies for Ansible
rich>=13.0.0
tabulate>=0.9.0
pyyaml>=6.0
jinja2>=3.1.0
cryptography>=41.0.0
paramiko>=3.0.0
